- WL#6965: Truncate UNDO logs.   - Increased the limit of innodb_max_undo_log_size from LONGLONG_MAX to     ULONGLONG_MAX.

+++ storage/innobase/srv/srv0srv.cc
@@ -118,7 +118,7 @@ for truncate (action is never aborted). */
 my_bool	srv_undo_truncate = FALSE;
 
 /** Maximum size of undo tablespace. */
-ulong	srv_max_undo_log_size;
+unsigned long long	srv_max_undo_log_size;
 
 /** UNDO logs that are not redo logged.
 These logs reside in the temp tablespace.*/
